LibWeb: Extract CanvasDrawImage class from CRC2D
